+++ dyatel/base/element.py
@@ -4,9 +4,11 @@ import time
 from logging import info
 from typing import Any
 
-from dyatel.base.driver import Driver
-from dyatel.dyatel_play.play_driver import PlayDriver
-from dyatel.dyatel_sel.core.core_driver import CoreDriver
+from playwright.sync_api import Page as PlaywrightDriver
+from appium.webdriver.webdriver import WebDriver as AppiumDriver
+from selenium.webdriver.remote.webdriver import WebDriver as SeleniumDriver
+
+from dyatel.base.driver_wrapper import DriverWrapper
 from dyatel.dyatel_play.play_element import PlayElement
 from dyatel.dyatel_sel.elements.mobile_element import MobileElement
 from dyatel.dyatel_sel.elements.web_element import WebElement
@@ -32,28 +34,31 @@ class Element(WebElement, MobileElement, PlayElement):
         self.name = name
         self.parent = parent
         self.wait = wait
-        self._initialized = False
-        self._driver_instance = Driver
 
-        self.element_class = self.__get_element_class()
+        self._initialized = False
+        self._driver_instance = DriverWrapper
+
+        self.element_class = self.__set_base_class()
         if self.element_class:
             super().__init__(locator=locator, locator_type=locator_type, name=name, parent=parent, wait=wait)
 
-    def __get_element_class(self):
+    def __set_base_class(self):
         """
         Get element class in according to current driver, and set him as base class
 
         :return: element class
         """
-        if PlayDriver.driver:
+        if isinstance(self.driver, PlaywrightDriver):
             Element.__bases__ = PlayElement,
             return PlayElement
-        elif CoreDriver.driver and CoreDriver.mobile:
+        elif isinstance(self.driver, AppiumDriver):
             Element.__bases__ = MobileElement,
             return MobileElement
-        elif CoreDriver.driver and not CoreDriver.mobile:
+        elif isinstance(self.driver, SeleniumDriver):
             Element.__bases__ = WebElement,
             return WebElement
+
+        # No exception due to delayed initialization
 
     # Following methods works same for both Selenium/Appium and Playwright APIs