+++ dyatel/base/page.py
@@ -2,11 +2,13 @@ from __future__ import annotations
 
 from logging import debug, info
 
-from dyatel.base.driver import Driver
+from playwright.sync_api import Page as PlaywrightDriver
+from appium.webdriver.webdriver import WebDriver as AppiumDriver
+from selenium.webdriver.remote.webdriver import WebDriver as SeleniumDriver
+
+from dyatel.base.driver_wrapper import DriverWrapper
 from dyatel.base.element import Element
-from dyatel.dyatel_play.play_driver import PlayDriver
 from dyatel.dyatel_play.play_page import PlayPage
-from dyatel.dyatel_sel.core.core_driver import CoreDriver
 from dyatel.dyatel_sel.pages.mobile_page import MobilePage
 from dyatel.dyatel_sel.pages.web_page import WebPage
 from dyatel.mixins.internal_utils import WAIT_PAGE
@@ -24,12 +26,12 @@ class Page(WebPage, MobilePage, PlayPage):
         :param name: name of page (will be attached to logs)
         :param driver_wrapper: set custom driver for page and page elements
         """
-        self._driver_instance = driver_wrapper if driver_wrapper else Driver
-        self.__set_page_class()
+        self._driver_instance = driver_wrapper if driver_wrapper else DriverWrapper
+        self.__set_base_class()
         super().__init__(locator=locator, locator_type=locator_type, name=name)
 
         if driver_wrapper:
-            if isinstance(driver_wrapper, Driver):
+            if isinstance(driver_wrapper, DriverWrapper):
                 self.set_driver(driver_wrapper)
             else:
                 self.set_driver(driver_wrapper.driver_wrapper)
@@ -110,19 +112,19 @@ class Page(WebPage, MobilePage, PlayPage):
         self._set_driver(driver_wrapper, Element)
         return self
 
-    def __set_page_class(self):
+    def __set_base_class(self):
         """
         Get page class in according to current driver, and set him as base class
 
         :return: page class
         """
-        if PlayDriver.driver:
+        if isinstance(self.driver, PlaywrightDriver):
             Page.__bases__ = PlayPage,
             return PlayPage
-        elif CoreDriver.driver and CoreDriver.mobile:
+        elif isinstance(self.driver, AppiumDriver):
             Page.__bases__ = MobilePage,
             return MobilePage
-        elif CoreDriver.driver and not CoreDriver.mobile:
+        elif isinstance(self.driver, SeleniumDriver):
             Page.__bases__ = WebPage,
             return WebPage
         else: