+++ dyatel/dyatel_sel/elements/mobile_element.py
@@ -3,6 +3,9 @@ from __future__ import annotations
 from logging import info
 from typing import Union, List, BinaryIO, Any
 
+from appium.webdriver.common.touch_action import TouchAction
+from selenium.webdriver.common.by import By
+
 from dyatel.dyatel_sel.core.core_driver import CoreDriver
 from dyatel.dyatel_sel.core.core_element import CoreElement
 from dyatel.dyatel_sel.sel_utils import get_legacy_selector, get_locator_type
@@ -24,6 +27,10 @@ class MobileElement(CoreElement):
         :param wait: include wait/checking of element in wait_page_loaded/is_page_opened methods of Page
         """
         self.is_safari_driver = CoreDriver.is_safari_driver
+        self.is_ios = CoreDriver.is_ios
+
+        self.top_bar_height = None
+        self.bottom_bar_height = None
 
         self.locator_type = locator_type if locator_type else get_locator_type(locator)
         self.locator, self.locator_type = get_legacy_selector(locator, self.locator_type)
@@ -33,7 +40,7 @@ class MobileElement(CoreElement):
     @property
     def all_elements(self) -> List[Any]:
         """
-        Get all wrapped elements with selenium bases
+        Get all wrapped elements with appium bases
 
         :return: list of wrapped objects
         """
@@ -50,7 +57,7 @@ class MobileElement(CoreElement):
         :return: self
         """
         if self.is_safari_driver:
-            self.wait_availability(timeout=timeout, silent=silent)
+            self.wait_availability(timeout=timeout, silent=True)
         else:
             super().wait_element(timeout=timeout, silent=silent)
 
@@ -96,15 +103,27 @@ class MobileElement(CoreElement):
 
         :return: self
         """
-        info(f'Hover over {self.name}')
-        self._action_chains\
-            .move_to_element(self.element)\
-            .move_by_offset(1, 1)\
-            .move_to_element(self.element)\
-            .perform()
+        self.wait_element(silent=True)
+
+        info(f'Tap to "{self.name}"')
+
+        if self.is_ios:
+            x, y = self.element.location.values()
+            y += self._get_top_bar_height()
+            TouchAction(self.driver).tap(x=x, y=y).perform()
+        else:
+            self._action_chains.click(on_element=self.element).perform()
         return self
 
-    def click_outside(self, x: int = 0, y: int = -5) -> MobileElement:  # TODO: Rework with `TouchAction`
+    def hover_outside(self, x: int = -100, y: int = -100) -> MobileElement:
+        """
+        Hover outside from current element
+
+        :return: self
+        """
+        return self.click_outside(x, y)
+
+    def click_outside(self, x: int = 0, y: int = -5) -> MobileElement:
         """
         Click outside of element. By default, 5px above  of element
 
@@ -113,11 +132,19 @@ class MobileElement(CoreElement):
         :return: self
         """
         self.wait_element(silent=True)
-        dx, dy = calculate_coordinate_to_click(self, x, y)
-        self._action_chains\
-            .move_to_element_with_offset(self.element, dx, dy)\
-            .click()\
-            .perform()
+
+        info(f'Tap outside from "{self.name}"')
+
+        if self.is_ios:
+            el_x, el_y = self.element.location.values()
+            el_y += self._get_top_bar_height()
+            TouchAction(self.driver).tap(x=el_x + x, y=el_y + y).perform()
+        else:
+            dx, dy = calculate_coordinate_to_click(self, x, y)
+            self._action_chains\
+                .move_to_element_with_offset(self.element, dx, dy)\
+                .click()\
+                .perform()
         return self
 
     def get_screenshot(self, filename: str, legacy: bool = True) -> BinaryIO:
@@ -142,7 +169,7 @@ class MobileElement(CoreElement):
 
     def _element_box(self) -> tuple:
         """
-        Get element coordinates on screen
+        Get element coordinates on screen for ios safari
 
         :return: element coordinates on screen (start_x, start_y, end_x, end_y)
         """
@@ -152,16 +179,60 @@ class MobileElement(CoreElement):
         start_x, start_y = el_location.values()
         h, w = self.element.size.values()
 
-        inner_height = self.driver.execute_script('return window.innerHeight')
-        outer_height = self.driver.execute_script('return window.outerHeight')
-        bars_size = outer_height - inner_height
+        if self.is_safari_driver:
+            inner_height = self.driver.execute_script('return window.innerHeight')
+            outer_height = self.driver.execute_script('return window.outerHeight')
+            bars_size = outer_height - inner_height
 
-        if bars_size > 110:  # FIXME: magick value
-            bar_size = bars_size / 4  # top and bottom bar shown
+            if bars_size > 110:  # There is no way to get top/bottom bars of safari with SafariDriver
+                bar_size = bars_size / 4  # top and bottom bar shown
+            else:
+                bar_size = bars_size / 2  # top bar shown, bottom hidden
         else:
-            bar_size = bars_size / 2  # top bar shown
+            bar_size = self._get_top_bar_height()
 
         if bar_size:
             start_y += bar_size
 
         return start_x, start_y, start_x+w, start_y+h
+
+    def _get_top_bar_height(self) -> int:
+        """
+        iOS only: Get top bar height
+
+        :return: self
+        """
+        if not self.top_bar_height:
+            self.driver_wrapper.switch_to_native()
+
+            top_bar = self.driver.find_element(
+                By.XPATH,
+                '//*[contains(@name, "SafariWindow")]/XCUIElementTypeOther[1]/XCUIElementTypeOther/XCUIElementTypeOther'
+            )
+            top_bar_height = top_bar.size['height']
+
+            self.driver_wrapper.switch_to_web()
+            return top_bar_height
+        else:
+            return self.top_bar_height
+
+    def _get_bottom_bar_height(self, force: bool = False) -> int:
+        """
+        iOS only: Get bottom bar height
+
+        :param force: get the new value forcly
+        :return: self
+        """
+        if force or not self.top_bar_height:
+            self.driver_wrapper.switch_to_native()
+
+            bottom_bar = self.driver.find_element(
+                By.XPATH,
+                '//*[@name="CapsuleViewController"]/XCUIElementTypeOther[1]'
+            )
+            bottom_bar_height = bottom_bar.size['height']
+
+            self.driver_wrapper.switch_to_web()
+            return bottom_bar_height
+        else:
+            return self.top_bar_height