+++ tests/mobile_tests/test_swipe.py
@@ -3,10 +3,10 @@ def get_scroll_position(driver):
 
 
 def test_swipe(second_playground_page, driver_wrapper):
-    expected_scroll = 392 if driver_wrapper.is_android else 1132
+    expected_scroll = range(300, 400) if driver_wrapper.is_android else range(1050, 1200)
     second_playground_page.swipe(0, 500, 0, 100, sleep=0.3)
     scroll = get_scroll_position(driver_wrapper.driver)
-    assert scroll == expected_scroll
+    assert scroll in expected_scroll
     second_playground_page.swipe(0, 100, 0, 500, sleep=0.3)
     scroll = get_scroll_position(driver_wrapper.driver)
     assert scroll == 0